restructure the logic
